Small fixes

+++ src/memory.c
@@ -162,7 +162,7 @@ static void freeObject(Obj *object) {
     }
     case OBJ_STRING: {
         ObjString *string = (ObjString *) object;
-        FREE_ARRAY(char, string->chars, string->length + 1);
+        FREE_ARRAY(char, string->chars, (size_t) string->length + 1);
         FREE(ObjString, string);
         break;
     }
